+++ mycodo/outputs/on_off_gpio.py
@@ -78,7 +78,7 @@ channels_dict = {
 # Output information
 OUTPUT_INFORMATION = {
     'output_name_unique': 'wired',
-    'output_name': "{}: GPIO".format(lazy_gettext('On/Off')),
+    'output_name': "GPIO: {}".format(lazy_gettext('On/Off')),
     'output_library': 'RPi.GPIO',
     'measurements_dict': measurements_dict,
     'channels_dict': channels_dict,